Rename write.[ch] to file.[ch]
